pyproject.toml
[tool.poetry]
name = "dz-27"
version = "0.1.0"
description = ""
authors = ["Chekhlenkov <113024470+Chekhlenkov@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dz_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
